Altered ErrorMethod to be an object and fixed rebasing issues.
