+++ src/main/kotlin/com/manimdsl/ManimParserVistor.kt
@@ -8,8 +8,6 @@ class ManimParserVisitor: ManimParserBaseVisitor<ASTNode>() {
     val currentSymbolTable = SymbolTableNode()
     private val semanticAnalyser = SemanticAnalysis()
 
-    private val finalDSLCode = mutableListOf<String>()
-
     override fun visitProgram(ctx: ManimParser.ProgramContext): ProgramNode {
         return ProgramNode(ctx.stat().map { visit(it) as StatementNode })
     }
@@ -90,7 +88,7 @@ class ManimParserVisitor: ManimParserBaseVisitor<ASTNode>() {
             method
 
         } else {
-            ErrorMethod()
+            ErrorMethod
         }
 
         return MethodCallNode(ctx.start.line, ctx.IDENT(0).symbol.text, dataStructureMethod, arguments)
