+++ src/test/kotlin/com/manimdsl/ASTExecutorTests.kt
@@ -15,7 +15,7 @@ class ASTExecutorTests {
                 "# code comment\n" +
                 "let y: Stack = new Stack;\n"
 
-        val (abstractSyntaxTree, symbolTable) = buildAST(program)
+        val (_, abstractSyntaxTree, symbolTable) = buildAST(program)
 
         val executor = ASTExecutor(abstractSyntaxTree, symbolTable, program.split("\n"))
 
@@ -55,7 +55,7 @@ class ASTExecutorTests {
     }
 
     // Assumes syntactically correct program
-    private fun buildAST(program: String): Pair<ProgramNode, SymbolTable> {
+    private fun buildAST(program: String): Triple<ExitStatus, ProgramNode, SymbolTable> {
         val parser = ManimDSLParser(program.byteInputStream())
         return parser.convertToAst(parser.parseFile().second)
     }
