+++ src/test/testFiles/valid/example.manimdsl
@@ -1,5 +1,4 @@
 let x = 1;
 let y: Stack = new Stack;
 y.push(x);
-x.push(2);
 let z = y.pop();
